requirements.txt
emmett>=2.5.5
psycopg2-binary>=2.9.9
pandas>=2.1.1
pycases>=0.1.1
result>=0.13.1